+++ package.json
@@ -14,8 +14,8 @@
     "browserify": "^12.0.0",
     "hihat": "^2.5.0",
     "minify": "^1.4.11",
-    "snazzy": "^1.0.1",
-    "standard": "^4.5.4",
+    "snazzy": "^2.0.1",
+    "standard": "^5.4.1",
     "tap-dev-tool": "^1.3.0",
     "tape": "^4.2.2",
     "trigger-event": "^1.0.0",
@@ -23,11 +23,11 @@
   },
   "dependencies": {
     "component-each": "^0.2.6",
-    "component-raf": "^1.2.0",
     "dift": "^0.1.2",
     "is-dom": "^1.0.5",
     "is-svg-attribute": "^1.0.2",
     "is-svg-element": "^1.0.1",
-    "setify": "^1.0.3"
+    "setify": "^1.0.3",
+    "union-type": "^0.1.6"
   }
 }
