+++ test/_diffNode.js
@@ -1,4 +1,6 @@
-import test from 'tape'
-import {diffNode} from '../src/diff'
-import * as actions from '../src/actions'
-import h from '../src/element'
+// import test from 'tape'
+// import {diffNode} from '../src/diff'
+// import * as actions from '../src/actions'
+// import h from '../src/element'
+//
+// test('diffNode', t => {})
